requirements.txt
django==4.0.2
djangorestframework==3.13.1
djangorestframework_simplejwt==5.0.0
channels==3.0.4